requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenSearch Client
opensearch-py==2.4.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Data Generation
faker==20.1.0

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Development
tabulate==0.9.0
